minor refactor of charities model and game model + minor changes to game results blade
